refactor: eliminate cfg.ts to break cycles
Split cfg.ts into two files to break the cfg.ts-rooted cycles madge
reports:
- util/config.ts: GlobalConfiguration, Analytics, Page (leaf, depends
  only on plugins/vfile.ts for ValidDateType)
- util/ctx.ts: QuartzConfig (co-located with BuildCtx)

Updated all 30 call sites to import from the new locations; removed
cfg.ts. Remaining cycles (util/ctx.ts <-> plugins/types.ts <->
components/types.ts) are type-only and structurally intrinsic to the
plugin API: plugin functions take BuildCtx as a parameter, and
QuartzConfig.plugins references PluginTypes. TypeScript handles these
fine at compile time.

Reduced cycles reported by madge from 2 cfg-rooted cycles (plus SCSS
false-positive) to 2 type-only cycles between ctx/plugin/component
type modules.
