+++ quartz/components/ContentMeta.tsx
@@ -4,8 +4,8 @@ import type { JSX } from "preact"
 import React from "react"
 import readingTime from "reading-time"
 
-import { type GlobalConfiguration } from "../cfg"
 import { type QuartzPluginData } from "../plugins/vfile"
+import { type GlobalConfiguration } from "../util/config"
 import { Backlinks } from "./Backlinks"
 import { formatTitle } from "./component_utils"
 import { DateElement } from "./Date"
