+++ quartz/components/PageList.tsx
@@ -3,8 +3,8 @@ import { toJsxRuntime, type Options } from "hast-util-to-jsx-runtime"
 import { h } from "hastscript"
 import { Fragment, jsx, jsxs } from "preact/jsx-runtime"
 
-import { type GlobalConfiguration } from "../cfg"
 import { type QuartzPluginData } from "../plugins/vfile"
+import { type GlobalConfiguration } from "../util/config"
 import { type FullSlug, resolveRelative } from "../util/path"
 import { formatTitle } from "./component_utils"
 import { locale } from "./constants"
