+++ quartz/plugins/transformers/tests/tablecaption.test.ts
@@ -5,7 +5,7 @@ import { h } from "hastscript"
 
 import type { BuildCtx } from "../../../util/ctx"
 
-import { QuartzConfig } from "../../../cfg"
+import { QuartzConfig } from "../../../util/ctx"
 import {
   TableCaption,
   isTextNode,
